🚚 clean up atom.io/react
